+++ lib/karafka/errors.rb
@@ -46,6 +46,9 @@ module Karafka
     ExpiredLicenseTokenError = Class.new(BaseError)
 
     # This should never happen. Please open an issue if it does.
-    InvalidCoordinatorState = Class.new(BaseError)
+    InvalidCoordinatorStateError = Class.new(BaseError)
+
+    # This should never happen. Please open an issue if it does.
+    StrategyNotFoundError = Class.new(BaseError)
   end
 end