+++ spec/lib/karafka/processing/strategies/default_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  it { expect(described_class::FEATURES).to eq([]) }
+end